Problem: rust->lurk impossible with wasm2wat incompatibility
Elaboration:

 - We don't support naked instructions in Wasm.lean (because it's insane
   to mix S-expressions with non-S-expressions! And during the
   requirement-gathering, we have collectively determined that for the
   purposes, it will be enough to only support operations).
 - Since wasm2wat generates WAT, it doesn't generate folded instructions
   by default, and for instructions that aren't foldable (like
   `i32.const 42`), it doesn't wrap them into parentheses.
 - In order for us to have an e2e demo of running Rust on Lurk, we need
   rustc to eventually emit WAT in the format we support.

Solution:
 - To do that, we need to wrap all the instructions into parens.
 - We have to, however, not touch well-formed S-exprs in WAT (for
   example type defs).
 - Thus, we have modelled WAT structures as a recursive data type `T`:

    ```
    <label : Str, prefix : [(Str | T)], instructions : [(Str | T)]> : T
    ```

 - In this triple, only `instructions` will be post-processed by:

    1. Wrapping normal instructions in "(" and ")".
    2. Changing "scope" instructions "s" into "(s" and changing "end"
       instructions into ")".

The logic by which we determine where the prefix ends is very simple. We
look ahead searching for an S-expression. If there is none, it means
that we're in instructions section and should append to instructions,
otherwise we're still parsing prefix.

Note that this approach allows us to parse **both module and
functions**!

Indeed, since elements of prefix and instructions are sum types over
string or T, as we parse module, we'll create a bunch of string prefix
entries for funcs. But each of these entries also parses into T itself,
creating lists of instructions, which is exactly our end goal.

Once we have the list of instructions we apply transformation described
above (it's not implemented yet, will be implemented in the next commit)
et viola.